move stdlib extensions into `[...].internal.stdlib`
