+++ docusync-gradle-plugin/src/main/kotlin/com/rickbusarow/docusync/internal/stdlib/file.kt
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-package com.rickbusarow.docusync.internal
+package com.rickbusarow.docusync.internal.stdlib
 
 import java.io.File
