+++ project.dic
@@ -25,6 +25,7 @@ kotest
 ktlint
 mainwave
 markwhitaker
+mkdirs
 modulecheck
 onboarding
 oneof
